Add about directory, add v1.3.0 release notes
