+++ mcweb/frontend/src/features/header/UserMenu.jsx
@@ -59,6 +59,9 @@ const UserMenu = () => {
           <MenuItem onClick={handleLogout}>
             Logout
           </MenuItem>
+          <MenuItem component={NavLink} to="release-notes" onClick={handleCloseUserMenu}>
+            Release Notes
+          </MenuItem>
         </Menu>
       </>
     );
